refactor: remove chat and video aggregates, separate files for chat and room domain objects
