--- src/aggregates/video/videoController.ts
@@ -1,1 +1,0 @@
-export interface VideoController {}
