+++ src/model/room/room.ts
@@ -1,68 +1,14 @@
-import { ChatReactions } from '../view/reactions/chatReactions'
-import { RoomReactions } from '../view/reactions/roomReactions'
-import { Entity, Repository } from './entity'
-import { Message, MessageContent, Notification, NotificationMessage, TextMessage } from './message'
+import { RoomReactions } from '../../view/reactions/roomReactions'
+import { Entity, Pair, Repository } from '../entity'
+import { Message, MessageContent, TextMessage } from '../message'
+import { Chat } from './chat'
 import { User, UserRepository } from './user'
-
-export interface Chat {
-  userJoined(user: User, chatReactions: ChatReactions): void
-
-  userLeft(user: User, chatReactions: ChatReactions): void
-
-  sendMessage(message: TextMessage, chatReactions: ChatReactions): void
-
-  get getMessages(): TextMessage[]
-}
-
-export class ChatImpl implements Chat {
-  private readonly messages: TextMessage[]
-
-  constructor() {
-    this.messages = []
-  }
-
-  userJoined(user: User, chatReactions: ChatReactions): void {
-    chatReactions.sendNotificationToRoom(new NotificationMessage(user, Notification.JOINROOM))
-    chatReactions.emitTextMessagesToClient(...this.messages)
-  }
-
-  userLeft(user: User, chatReactions: ChatReactions): void {
-    chatReactions.sendNotificationToRoom(new NotificationMessage(user, Notification.LEAVEROOM))
-  }
-
-  sendMessage(message: TextMessage, chatReactions: ChatReactions): void {
-    this.messages.push(message)
-    chatReactions.sendTextMessagesToRoom(message)
-  }
-
-  get getMessages(): TextMessage[] {
-    return this.messages
-  }
-}
 
 export class RoomId {
   roomName: string
 
   constructor(roomName: string) {
     this.roomName = roomName
-  }
-}
-
-export class Pair<X, Y> {
-  private readonly x: X
-  private readonly y: Y
-
-  constructor(x: X, y: Y) {
-    this.x = x
-    this.y = y
-  }
-
-  get getX(): X {
-    return this.x
-  }
-
-  get getY(): Y {
-    return this.y
   }
 }
